Apple/pyproject.toml
[tool.poetry]
name = "apple"
version = "0.1.0"
description = ""
authors = ["TonyChou <cool21540125@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
load-dotenv = "^0.1.0"
python-dotenv = "^1.0.0"
pyjwt = { extras = ["crypto"], version = "^2.8.0" }


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
